keep data parity with content types.

+++ app/controllers/wh/content/all-types.js
@@ -1,8 +1,16 @@
+import getItemModelName from 'appkit/utils/model';
+
 export default Ember.ArrayController.extend({
   actions: {
     deleteType: function (contentType) {
-      this.removeObject(contentType);
-      contentType.destroyRecord();
+      if (window.confirm('Are you sure? Confirm to delete this type and all associated data.')) {
+        // remove all associated data from Firebase
+        window.ENV.firebase.child('data').child(contentType.get('id')).remove();
+
+        // remove content type
+        contentType.destroyRecord();
+      }
+
     },
     gotoEdit: function (name) {
       this.transitionToRoute('form', name);
